Support for kubeadm bootstrapping & using new imgs
This patch adds support for using kubeadm for bootstrapping onto the new
CAPV images. This is to align CAPV's features with those in CAPA.

Add unit test for GetMachineRole

This patch adds a unit test for GetMachineRole.

Updated CAPI to same version as CAPA

This patch updates the CAPI dependency to the same version used by CAPA.

Update CI provider component logging

This patch sets the log level in the CI provider components file to 6.

Remove unused code in GoVmomi utils

This patch removes unused code in GoVmomit utils.

Centralize API updates

This patch removes the code that updated the taskref from inside the
GoVmomi create callstack. The machines are now only updated as part of a
deferred routine in the actuators.

Return if find VM by moref

This patch returns from the Create call if the VM is found by its MoRef.

Refactored to use Cluster & Machine Contexts

Akin to CAPA's Scopes, this patch introduces support for Cluster &
Machine contexts, making writing code much, much easier, as well as
centralizing all of the boilerplate work that happens at the beginning
and end of each request into the context.

Initialize loggers with actuator IDs

This patch ensures the cluster and machine actuator loggers have the
correct prefix.

Lock the klog library to v0.2.0

Updating the project with support for klogr caused the klog dependency
to be updated to v0.3.1, and that was causing CI to break.

Support disabling validation on gen yaml

This patch adds support for disabling the validation of certain env vars
when generating yaml with generate-yaml.sh

Use CAPI 0.1.3 image

This patch switches to using the CAPI 0.1.3 image due to an error
starting CAPI inside of Kind with the 0.1.2 image. Please see
https://github.com/kubernetes-sigs/cluster-api/issues/999 for more
information.

Refactored Actuators, removed informers

This patch refactors the actuators into discreet packages and removed
the use of informers. This patch also relocates the setup code for the
controller into the manager binary's main file.

Update e2e kubectl bin to 1.14.2

This patch updates the kubectl binary used by the e2e job to 1.14.2.

Add SSH auth keys to cloud-init

This patch forwards the ClusterConfig's SSH auth keys list into the
cloud-init data.

Allow a machine to return control plane endpoint

This patch allows a machine to return a control plane endpoint to
satisfy the ProviderDeployer interface. Please see the Slack
conversation
https://kubernetes.slack.com/archives/C8TSNPY4T/p1560696064368600 for
more information.